refactor(NcAppNavigationToggle): move to `NcAppNavigation` to prevent public export
