--- src/components/NcAppNavigationToggle/index.ts
@@ -1,6 +1,0 @@
-/**
- * SPDX-FileCopyrightText: 2019 Nextcloud GmbH and Nextcloud contributors
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-export { default } from './NcAppNavigationToggle.vue'
